fix: change parameters type from int to double

+++ JavaFiapTokio2024/src/logica/exercicios/aula11/Exercicio01.java
@@ -6,7 +6,7 @@ public class Exercicio01 {
 		System.out.printf("Area do terreno de dimensoes %.2f x %.2f = %.2f", 2f, 5f, calcularArea(2, 5));
 	}
 
-	static double calcularArea(int largura, int comprimento) {
+	static double calcularArea(double largura, double comprimento) {
 		return largura * comprimento;
 	}
 }
